backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.1
